moved main activity and fragments into ui package
