+++ app/src/main/java/com/lambdaschool/hackathon_portal/di/AppComponent.kt
@@ -2,7 +2,7 @@ package com.lambdaschool.hackathon_portal.di
 
 import android.app.Activity
 import android.app.Application
-import com.lambdaschool.hackathon_portal.fragments.dashboard.DashboardFragment
+import com.lambdaschool.hackathon_portal.ui.fragments.dashboard.DashboardFragment
 import dagger.BindsInstance
 import dagger.Component
 import javax.inject.Singleton
