--- app/src/main/java/com/lambdaschool/hackathon_portal/fragments/login/LoginViewModel.kt
@@ -1,1 +1,0 @@
-package com.lambdaschool.hackathon_portal.fragments.login
