+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/loginFragment"
-        android:name="com.lambdaschool.hackathon_portal.fragments.login.LoginFragment"
+        android:name="com.lambdaschool.hackathon_portal.ui.fragments.login.LoginFragment"
         android:label="fragment_login"
         tools:layout="@layout/fragment_login" >
         <action
@@ -16,7 +16,7 @@
     </fragment>
     <fragment
         android:id="@+id/dashboardFragment"
-        android:name="com.lambdaschool.hackathon_portal.fragments.dashboard.DashboardFragment"
+        android:name="com.lambdaschool.hackathon_portal.ui.fragments.dashboard.DashboardFragment"
         android:label="fragment_dashboard"
         tools:layout="@layout/fragment_dashboard" >
         <action
